Rete

Documentation for Rete.

Facts

The facts in our reasoning system can be arbitrary Julia objects. It's best to restrict facts to immutable objects though so that they can't be altered once they're stored in the network or conclusions have been made.

The Network

The reasoning is performed by a network of nodes. Facts flow through the network from node to node.

Each node has a set of input nodes (those that send it facts), and a set of outputs (those nodes to which it sends facts). connect is used to construct the network by linking nodes together.

Some nodes filter the facts and only pass through those that satisfy some predicate or are instances of a certain type.

Some nodes store facts.

Join nodes have two input streams. A function is applied to all possible combinations of facts coming in from the two streams. The function can call emit to assert a new fact to the network.

That's the theory. In practice, its simpler if a given node performs more than one of these roles. One such example is IsaMemoryNode, which filters facts that match a type parameter and remember only those facts.

Flow of Facts through the Network

A node receives a fact via its receive method.

A node distributes a fact to its outputs using its emit method, which calls receive for each of the node's outputs.

Index

Rete.AbstractMemoryNode

Rete.AbstractReteJoinNode

Rete.AbstractReteNode

Rete.IsaMemoryNode

Rete.JoinNode

Rete.connect

Rete.connect_a

Rete.connect_b

Rete.emit

Rete.ensure_IsaMemoryNode

Rete.find_memory_for_type

Rete.inputs

Rete.label

Rete.outputs

Rete.receive

Glossary

Rete.AbstractMemoryNode — Type

AbstractMemoryNode is the abstract supertype of all Rete memory nodes.

Rete.AbstractReteJoinNode — Type

AbstractReteJoinNode is the abstract supertype of all Rete join nodes.

Rete.AbstractReteNode — Type

AbstractReteNode is the abstract supertype of all Rete nodes.

Rete.IsaMemoryNode — Type

IsaMemoryNode is a type of memory node that only stores facts of the specified type. Facts of other types are ignored.

Rete.JoinNode — Type

JoinNode implements a join operation between two streams of inputs A abd B. The first argiment of join_function is the JoinNode itself. The second comes from the A stream. The third comes from the B stream. join_function should call emit for each new fact it wants to assert.

Rete.connect — Method

connect(from, to)

makes to an output of from and from an input of to.

Rete.connect_a — Method

connect_a(from::AbstractMemoryNode, to::JoinNode)

Connect a memory node to the a input of a JoinNode.

Rete.connect_b — Method

connect_b(from::AbstractMemoryNode, to::JoinNode)

Connect a memory node to the b input of a JoinNode.

Rete.emit — Function

emit(node, fact)

Distribute's fact to each of node's outputs by calling receive on the output node and fact.

Rete.ensure_IsaMemoryNode — Method

ensure_IsaMemoryNode(root, typ::Type)::IsaTypeNode

Find the IsaMemoryNode for the specified type, or make one and add it to the network.

Rete.find_memory_for_type — Method

find_memory_for_type(root, typ::Type)::UnionPNothing, IsaTypeNode}

If there's a memory node in the Rete represented by root that stores objects of the specified type then return it. Otherwiaw return nothing.

Rete.inputs — Function

inputs(node)

Returns the inputs of node – those nodes which can send it facts.

Rete.label — Function

label(node)

Returns node's label.

Rete.outputs — Function

outputs(node)

Returns the outputs of node – those nodes to which it can send facts.

Rete.receive — Function

receive(node, fact)

receive is how node is given a new fact.