Rete

Documentation for Rete.

Rete currently implements memory for Julia types, forward chaining joins, and backward chaining filter and extrema operations.

Facts

The facts in our reasoning system can be arbitrary Julia objects. It's best to restrict facts to immutable objects though so that they can't be altered once they're stored in the network or conclusions have been made. There is no mechanism for retracting a conclusion if a fact is altered.

The Network

The reasoning is performed by a network of nodes. Facts flow through the network from node to node.

Each node has a set of input nodes (those that send it facts), and a set of outputs (those nodes to which it sends facts). connect is used to construct the network by linking nodes together.

Some nodes filter the facts and only pass through those that satisfy some predicate or are instances of a certain type.

Some nodes store facts.

Join nodes have multiple distinct input streams. A function is applied to all possible combinations of facts coming in from these streams. The function can call emit to assert a new fact to the network.

That's the theory. In practice, its simpler if a given node performs more than one of these roles. One such example is IsaMemoryNode, which filters facts that match a type parameter and remember only those facts.

We assume that the network is fully constructed before any facts are asserted. Adding inputs to a JoinNode doesn't cause existing facts from those inputs to be processed.

Layers

The network might best be constructed in layers. A single root node forms the top layer. It serves as the recipient of new facts. It distributes those facts to the next layer, which consists of memory nodes. A third layer consists of join nodes, typically defined by rules. The join nodes might conclude new facts which they pass to the root node.

Flow of Facts through the Network

A node receives a fact via its receive method.

A node distributes a fact to its outputs using its emit method, which calls receive for each of the node's outputs.

Rules

The @rule macro makes it easier to create rules and add them to a Rete.

As a contrived exanple, lets create a network that creates pairs of letters when the second letter in the pair is the next letter of the alphabet from the first.

using Rete

@rule PairConsectutiveLetters(a::Char, b::Char, ::String) begin
    if codepoint(a) + 1 == codepoint(b)
        emit(a * b)
    end
end

@rule will define a singleton type named PairConsectutiveLetters to represent the rule. @rule defines an install method that will add the rule to a network. The instance of PairConsectutiveLetters implements the join function of the JoinNode.

# Create the knowledgebase:
root = ReteRootNode("root")
install(root, PairConsectutiveLetters())

# Assert the characters 'a' through 'e' into the knowledgebase:
for c in 'a':'e'
    receive(root, c)
end

askc(Collector{String}(), root)

4-element Vector{String}:
 "ab"
 "bc"
 "cd"
 "de"

Backward Chaining

Rete currently provides limited support for backward chaining by using the BackwardFilterNode and BackwardExtremumNode node types.

There is not yet any facility to make it easier to integrate these nodes into the network. You will need to use the node constructors and connect to add them by hand.

Querying and Aggregation

The function askc can be used to query the network. askc takes either a continuation function or an Aggregator as its first argument.

askc also takes either a node that supports it, e.g. memory nodes or backweard chaining nodes; or a ReteRootNode representing your knowledge base, and a fact type.

The continuation finction is callled on each fact that askc finds.

If an aggregator is passed as the first argument then it will perform the specified aggregation over the course of the query and that call to askc will return the aggregation result.

The currenty supported aggregators are:

Counter

Collector

using Rete
kb = ReteRootNode("My Knowledge Base")
connect(kb, IsaMemoryNode{Int}())
receive.([kb], 1:5)
nothing

askc(Counter(), kb, Int)

5

askc(Collector{Int}(), kb)

5-element Vector{Int64}:
 1
 2
 3
 4
 5

Note that askc can infer its third argument from the Collector.

let
    sum = 0
    askc(kb, Int) do i
        sum += i
    end
    sum
end

15

Index

Rete.AbstractMemoryNode

Rete.AbstractReteJoinNode

Rete.AbstractReteNode

Rete.AbstractReteRootNode

Rete.Aggregator

Rete.BackwardChaining

Rete.BackwardExtremumNode

Rete.BackwardFilterNode

Rete.CanInstallRulesTrait

Rete.Collector

Rete.Counter

Rete.HasSetOfInputsTrait

Rete.HasSetOfOutputsTrait

Rete.IsaMemoryNode

Rete.JoinNode

Rete.ReteRootNode

Rete.Rule

Rete._add_input

Rete._add_output

Rete.add_forward_trigger

Rete.askc

Rete.askc

Rete.connect

Rete.copy_facts

Rete.emit

Rete.emits

Rete.ensure_memory_node

Rete.fact_count

Rete.find_memory_for_type

Rete.input_count

Rete.inputs

Rete.install

Rete.is_forward_trigger

Rete.is_memory_for_type

Rete.kb_counts

Rete.kb_stats

Rete.label

Rete.output_count

Rete.outputs

Rete.receive

Rete.walk_by_outputs

Rete.@rule

Glossary

Rete.AbstractMemoryNode — Type

AbstractMemoryNode is the abstract supertype of all Rete memory nodes.

Each concrete sybtype should implement is_memory_for_type.

Rete.AbstractReteJoinNode — Type

AbstractReteJoinNode is the abstract supertype of all Rete join nodes.

Rete.AbstractReteNode — Type

AbstractReteNode is the abstract supertype of all Rete nodes.

Rete.AbstractReteRootNode — Type

AbstractReteRootNode

AbstractReteRootNode is the abstract supertype for all root nodes of a Rete.

Rete.Aggregator — Type

Aggergator is the abstract supertype for all query aggregators.

Rete.BackwardChaining — Type

BackwardChaining is the abstract supertype for all backward chaining Rete nodes.

Rete.BackwardExtremumNode — Type

BackwardExtremumNode(comparison, extractor, label)

When askced, provides one value: the fact with the most extreme value (based on comparison) of extractor applied to each input fact at the time askc was called.

Rete.BackwardFilterNode — Type

 BackwardFilterNode(filter_function, label)

When askced, passes through from its inputs only those facts which satisfy predicate.

Rete.CanInstallRulesTrait — Type

CanInstallRulesTrait

Having this trait gives the root node of a knowledge base the install method to facilitate adding rules to the network.

You can add this trait to YourType with

CanInstallRulesTrait(::Type{<:YourType}) = CanInstallRulesTrait())

Rete.Collector — Type

Collector{T}()

returns an Aggregator that can be passed as the "continuation" function of askc so that askc will return a vector of the objects that it found.

Rete.Counter — Type

Counter()

returns an Aggregator that can be passed as the "continuation" function of askc so that askc will return the number of objects that it found.

Rete.HasSetOfInputsTrait — Type

HasSetOfInputsTrait(type)

A Rete node type with the HasSetOfInputsTrait stores its inputs as a set.

Any struct that is a subtype of AbstractReteNode with the field

    inputs::Set{AbstractReteNode}

will have this trait.

Rete.HasSetOfOutputsTrait — Type

HasSetOfOuputsTrait(type)

A Rete node type with the HasSetOfOutputsTrait stores its outputs as a set.

Any struct that is a subtype of AbstractReteNode with the field

    outputs::Set{AbstractReteNode}

will have this trait.

Rete.IsaMemoryNode — Type

IsaMemoryNode is a type of memory node that only stores facts of the specified type. Facts of other types are ignored.

Rete.JoinNode — Type

JoinNode implements a join operation between multiple streams of inputs. The first argiment of join_function is the JoinNode itself. The remaining arguments come from the input streams of the join node. join_function should call emit for each new fact it wants to assert.

Rete.ReteRootNode — Type

ReteRootNode

ReteRootNode serves as the root node of a Rete network.

If you need a specialized root node for your application, see AbstractReteRootNode and CanInstallRulesTrait.

Rete.Rule — Type

Rule is an abstract supertype for all rules.

Rete._add_input — Function

_add_input(from, to)

adds from as an input to to.

User code should never call _add_input directly, only through connect. Users might specialize _add_input if implementing a node that needs to specialize how it storesa its inputs.

Rete._add_output — Function

_add_output(from, to)

adds to as an output of from.

User code should never call _add_output directly, only through connect. Users might specialize _add_output if implementing a node that needs to specialize how it storesa its outputs.

Rete.add_forward_trigger — Method

add_forward_trigger(n::JoinNode, from::AbstractReteNode)

adds from as a forward trigger of the JoinNode.

When a JoinNode receives a fact from one of its forward trigger inputs, it joins that input with all combinations of facts from other inputs and performs its join_function. Otherwise the JoinNode is not triggered.

Rete.askc — Function

askc(continuation::Function, node)

Calls continuation on each fact available from node.

Rete.askc — Method

askc(continuation::Function, root::AbstractReteRootNode, t::Type)

calls continuation on every fact of the specified type (or its subtypes) that are stored in the network rooted at root.

Assumes all memory nodes are direct outputs of root.

Rete.connect — Method

connect(from::AbstractReteNode, to::AbstractReteNode)

makes to an output of from and from an input of to.

connect calls _add_input and _add_output to do this.

Rete.copy_facts — Method

copy_facts(from_kb::AbstractReteRootNode, to_kb::AbstractReteRootNode, fact_types)

Copues facts if the specified fact_type from from_kb to to_kb.

for multiple fact types, one can broadcast over a collection of fact types.

Rete.emit — Function

emit(node, fact)

Distribute's fact to each of node's outputs by calling receive on the output node and fact.

Rete.emits — Method

emits(rule::Type)

Returns a Tuple of the types which rule is declared to emit.

Rete.ensure_memory_node — Method

ensure_memory_node(root::AbstractReteRootNode, typ::Type)::IsaTypeNode

Find a memory node for the specified type, or make one and add it to the network.

The default is to make an IsaMemoryNode. Specialize this function for a Type to control what type of memory node should be used for that type.

Rete.fact_count — Method

fact_count(node)

For memory nodes, return the number of facts currently stored in the node's memory, otherwise return nothing.

Rete.find_memory_for_type — Method

find_memory_for_type(root, typ::Type)::Union(Nothing, AbstractMemoryNode}

If there's a memory node in the Rete represented by root that stores objects of the specified type then return it. Otherwise return nothing.

Rete.input_count — Method

input_count(node)

returns the number of inputs to the node.

Note that for join nodes, this is the number of parameters rather than the number of nodes that emit facts to the join.

Rete.inputs — Function

inputs(node)

Returns the inputs of node – those nodes which can send it facts – as an iterable collection.

Rete.install — Method

install(root, rule::Type)

Installs the rule or rule group into the Rete rooted at root.

Rete.is_forward_trigger — Method

is_forward_trigger(::JoinNode, from::AbstractReteNode)

returns true if from is a forward trigger input of the JoinNode.

Rete.is_memory_for_type — Method

is_memory_for_type(node, typ::Type)::Bool

returns true if node stores objects of the specified type.

Used by find_memory_for_type.

Rete.kb_counts — Method

kb_counts(root::AbstractReteRootNode)

Returns a Dict{Type, Int} of the number of facts of each type.

Rete.kb_stats — Method

kb_stats(io, root)

Show the input count, output count, fact count and label for each node.

Rete.label — Function

label(node)

Returns node's label.

Rete.output_count — Method

output_count(node)

returns the number of outputs from the node.

Rete.outputs — Function

outputs(node)

Returns the outputs of node – those nodes to which it can send facts – as an iterable collection.

Rete.receive — Function

receive(node, fact)

receive is how node is given a new fact.

An application calls receive on the root node to assert a new fact to the network.

Rete.walk_by_outputs — Method

walkbyoutputs(func, node::AbstractReteNode)

Walks the network rooted at `root', applying func to each node.

Rete.@rule — Macro

@rule Rulename(a::A_Type, b::B_Type, ...) begin ... end

Defines a rule named Rulename. A singleton type named Rulename will be defined to represent the rule. An install method is defined for that type which can be used to add the necessary nodes and connections to a Rete to implement the rule.

The default supertype of a rule struct is Rule. When it is desirable to group rules together, one can define an abstract type that is a type descendant of Rule and use that as a dotted prefix to RuleName. The RuleName in the @rule invocation is MyGroup.MyRule then the supertype of MyRule will be MyGroup, rather than Rule.

A rule can have arbitrarily many parameters. The parameter list can also include clauses with no variable name. Such clauses identify the types of facts that the rule might assert. Memory nodes for these types will be added to the Rete if not already present.

The first expression of the rule can be call-like expression of RULE_DECLARATIONS. Its "parameters" can be declarations of one of the forms

FORWARDTRIGGERS(argumentnames...)`

Only the inputs for the specified argument names will serve as forward triggers. For backward compatibility, if there is no RULE_DECLARATIONS expression then all inputs are forward triggers.

CUSTOM_INSTALL()`

No install method will be automatically generated. The developer must implement an install method for this rule.