Rete

Documentation for Rete.

In its current form, this package only supports forward chaining rules.

Facts

The facts in our reasoning system can be arbitrary Julia objects. It's best to restrict facts to immutable objects though so that they can't be altered once they're stored in the network or conclusions have been made.

The Network

The reasoning is performed by a network of nodes. Facts flow through the network from node to node.

Each node has a set of input nodes (those that send it facts), and a set of outputs (those nodes to which it sends facts). connect is used to construct the network by linking nodes together.

Some nodes filter the facts and only pass through those that satisfy some predicate or are instances of a certain type.

Some nodes store facts.

Join nodes have multiple distinct input streams. A function is applied to all possible combinations of facts coming in from these streams. The function can call emit to assert a new fact to the network.

That's the theory. In practice, its simpler if a given node performs more than one of these roles. One such example is IsaMemoryNode, which filters facts that match a type parameter and remember only those facts.

We assume that the network is fully constructed before any facts are asserted. Adding inputs to a JoinNode doesn't cause existing facts from those inputs to be processed.

Layers

The network might best be constructed in layers. A single root node forms the top layer. It serves as the recipient of new facts. It distributes those facts to the next layer, which consists of memory nodes. A third layer consists of join nodes, typically defined by rules. The join nodes might conclude new facts which they pass to the root node.

Flow of Facts through the Network

A node receives a fact via its receive method.

A node distributes a fact to its outputs using its emit method, which calls receive for each of the node's outputs.

Rules

The @rule macro makes it easier to create rules and add them to a Rete.

As a contried exanple, lets create a network that creates pairs of letters when the second letter in the pair is the next letter of the alphabet from the first.

using Rete

@rule PairConectutiveLetters(a::Char, b::Char, ::String) begin
    if codepoint(a) + 1 == codepoint(b)
        emit(a * b)
    end
end

@rule will define a singleton type named PairConectutiveLetters to represent the rule. @rule defines an install method that will add the rule to a network. The instance of PairConectutiveLetters implements the join function of the JoinNode.

# Create the knowledgebase:
root = ReteRootNode("root")
install(root, PairConectutiveLetters())

# Assert the characters 'a' through 'e' into the knowledgebase:
for c in 'a':'e'
    receive(root, c)
end

collecting() do c
    askc(c, root, String)
end

4-element Vector{Any}:
 "ab"
 "bc"
 "cd"
 "de"

Index

Rete.AbstractMemoryNode

Rete.AbstractReteJoinNode

Rete.AbstractReteNode

Rete.IsaMemoryNode

Rete.JoinNode

Rete.ReteRootNode

Rete.Rule

Rete.askc

Rete.collecting

Rete.connect

Rete.emit

Rete.ensure_IsaMemoryNode

Rete.find_memory_for_type

Rete.inputs

Rete.install

Rete.kb_counts

Rete.label

Rete.outputs

Rete.receive

Rete.@rule

Glossary

Rete.AbstractMemoryNode — Type

AbstractMemoryNode is the abstract supertype of all Rete memory nodes.

Rete.AbstractReteJoinNode — Type

AbstractReteJoinNode is the abstract supertype of all Rete join nodes.

Rete.AbstractReteNode — Type

AbstractReteNode is the abstract supertype of all Rete nodes.

Rete.IsaMemoryNode — Type

IsaMemoryNode is a type of memory node that only stores facts of the specified type. Facts of other types are ignored.

Rete.JoinNode — Type

JoinNode implements a join operation between multiple streams of inputs. The first argiment of join_function is the JoinNode itself. The remaining arguments come from the input streams of the join node. join_function should call emit for each new fact it wants to assert.

Rete.ReteRootNode — Type

ReteRootNode serves as the root node of a Rete network.

Rete.Rule — Type

Rule is an abstract supertype for all rules.

Rete.askc — Function

askc(continuation, node)

Calls continuation on each fact available from node.

Rete.collecting — Method

collecting(body)

runs body, passing it a continuation of one argument that collects the values it's called with. collecting returns those values.

Rete.connect — Method

connect(from, to)

makes to an output of from and from an input of to.

When to is a join node then a third parameter (a positive integer) identifies which parameter position from feed in to.

Rete.emit — Function

emit(node, fact)

Distribute's fact to each of node's outputs by calling receive on the output node and fact.

Rete.ensure_IsaMemoryNode — Method

ensure_IsaMemoryNode(root, typ::Type)::IsaTypeNode

Find the IsaMemoryNode for the specified type, or make one and add it to the network.

Rete.find_memory_for_type — Method

find_memory_for_type(root, typ::Type)::UnionPNothing, IsaTypeNode}

If there's a memory node in the Rete represented by root that stores objects of the specified type then return it. Otherwise return nothing.

Rete.inputs — Function

inputs(node)

Returns the inputs of node – those nodes which can send it facts.

Rete.install — Function

install(root::ReteRootNode, rule::Type)

Installs the rule or rule group into the Rete rooted at root.

Rete.kb_counts — Method

kb_counts(root::ReteRootNode)

Returns a Dict{Type, Int} of the number of facts of each type.

Rete.label — Function

label(node)

Returns node's label.

Rete.outputs — Function

outputs(node)

Returns the outputs of node – those nodes to which it can send facts.

Rete.receive — Function

receive(node, fact)

receive is how node is given a new fact.

An application calls receive on the root node to assert a new fact to the network.

Rete.@rule — Macro

@rule Rulename(a::A_Type, b::B_Type, ...) begin ... end

Defines a rule named Rulename. A singleton type named Rulename will be defined to represent the rule. An install method is defined for that type which can be used to add the necessary nodes and connections to a Rete to implement the rule.

The default supertype of a rule struct is Rule. When it is desirable to group rules together, one can define an abstract type that is a type descendant of Rule and use that as a dotted prefix to RuleName. The RuleName in the @rule invocation is MyGroup.MyRule then the supertype of MyRule will be MyGroup, rather than Rule.

A rule can have arbitrarily many parameters. The parameter list can also include clauses with no variable name. Such clauses identify the types of facts that the rule might assert. Memory nodes for these types will be added to the Rete if not already present.